+++ js/main.js
@@ -2,7 +2,6 @@ import { getNewCardElement } from './similar.js';
 import {generateData} from './data.js';
 import './user-form.js';
 import './map.js';
-import './slider.js';
 
 const data = generateData();
 const map = document.querySelector('#map-canvas');
